cleaning up everything a little bit and improve initialization component factory
