+++ src/main/webapp/js/widgets/popup/Popup.js
@@ -41,8 +41,6 @@ define(function (require) {
     var Widget = require('widgets/Widget');
     var $ = require('jquery');
     
-    var componentFactory = require('components/dev/form/ComponentFactory');
-    
     var React = require('react');
 
     /**
@@ -123,92 +121,9 @@ define(function (require) {
                 GEPPETTO.Console.log("Hooked up custom handlers for " + this.id);
             }
             
-            this.formStuff();
-
             return this;
         },
         
-        formStuff: function(callback){
-        	var formCallback = callback;
-        	
-        	var id = "gptForm";
-        	
-        	var schema = {
-      			  type: "object",
-      			  required: ["experimentName", "timeStep", "length", "simulator", "numberProcessors"],
-      			  properties: {
-      				experimentName: {type: "string", title: "Experiment Name"},
-      				timeStep:{type:'number', title: 'Time Step (s)'},
-      				length:{type:'number', title: 'Length (s)'},
-      				simulator:{
-	    			      type: "string",
-	    			      title: "Simulator",
-	    			      enum: ["neuronSimulator", "lemsSimulator", "neuronNSGSimulator"],
-	    			      enumNames: ["Neuron", "jLems", "Neuron on NSG"]
-      				},
-
-      				numberProcessors:{type:'number', title: 'Number of Processors'}
-      			  }
-      			};
-      	
-        	var formData= {
-        			experimentName: Project.getActiveExperiment().getName(),
-        			timeStep: Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getTimeStep(),
-        			length: Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getLength(),
-        			simulator:Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].getSimulator(),
-        			numberProcessors: 1
-        	};
-        	
-        	var submitHandler = function(){
-        		GEPPETTO.Flows.showSpotlightForRun(formCallback);
-        		formWidget.destroy();
-        	};
-        	
-        	var errorHandler = function(){
-        		
-        	};
-        	
-        	var changeHandler = function(formObject){
-        		for (var key in formObject.formData) {
-        			if (formObject.formData[key] != this.formData[key]){
-        				if (key == 'experimentName'){
-        					$("#experimentsOutput").find(".activeExperiment").find("td[field='name']").html(formObject.formData[key]).blur();
-        				}
-        				else if (key == 'timeStep'){
-        					$("#experimentsOutput").find(".activeExperiment").find("td[field='timeStep']").html(formObject.formData[key]).blur();
-        				}
-						else if (key == 'length'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='length']").html(formObject.formData[key]).blur();					
-						}
-						else if (key == 'simulator'){
-							Project.getActiveExperiment().simulatorConfigurations[window.Instances[0].getId()].setSimulatorParameter('numberProcessors', formObject.formData[key]).blur();
-						}
-						else if (key == 'numberProcessors'){
-							$("#experimentsOutput").find(".activeExperiment").find("td[field='simulatorId']").html(formObject.formData[key]).blur();
-						}
-        				this.formData[key] = formObject.formData[key];
-        			}
-        		}
-    		};
-      	
-//        	componentFactory.getComponent('FORM',
-//        			{schema:schema, formData:formData, submitHandler:submitHandler, errorHandler:errorHandler, changeHandler:changeHandler},
-//        			document.getElementById(this.id));
-
-//        	componentFactory.addComponent('FORM',
-//        			{id: "takaForm", name:"paka", schema:schema, formData:formData, submitHandler:submitHandler, errorHandler:errorHandler, changeHandler:changeHandler});
-        	
-    		
-        	var panelChildren = [];
-        	panelChildren.push(componentFactory.getComponent('DIVFEO',{id:"pp", text:'xaka'}));
-        	panelChildren.push(componentFactory.getComponent('FORM',{id: "pp2", name:"mm", schema:schema, formData:formData, submitHandler:submitHandler, errorHandler:errorHandler, changeHandler:changeHandler}));
-        	
-        	var panelComponent = componentFactory.addComponent('PANEL',	{id: "kklr", name:"pl"});
-        	panelComponent.addChildren(panelChildren);
-        	console.log(panelComponent);
-        	
-        },
-
         /**
          * Sets the message that is displayed inside the widget through an instance of type Text
          *
